fixed tiling region
