+++ views/layout.jade
@@ -9,6 +9,7 @@ html(lang='en')
         script(src='/js/html5shiv.js')
         script(src='/js/respond.min.js')
         script(type='text/javascript', src ="js/pixi.js")
+        script(type='text/javascript', src ="js/linkedlist.js")
         script(type='text/javascript', src ="js/game/MapParts.js")
         script(type='text/javascript', src ="js/game/StartButton.js")
         script(type='text/javascript', src ="js/gameplay.js")
